Conflicts:
	.gitignore

+++ Makefile
@@ -21,9 +21,11 @@ COMPILE_OPTIONS=$(WARN) -d $(OBJDIR) -sourcepath $(SRCDIR) -classpath $(CLASSPAT
 all: compile
 
 run: compile
-	java $(RUN_OPTIONS)
+	java $(RUN_OPTIONS) 
+
 run_extraction: compile
 	java $(RUN_EXTRACTION)
+
 compile: $(OBJDIR) $(OBJ)
 
 $(OBJDIR)/%.class: $(SRCDIR)/%.java
@@ -33,9 +35,11 @@ $(OBJDIR):
 	@test -d $(OBJDIR) || mkdir $(OBJDIR)
 
 clean:
-	rm -fr $(OBJDIR)
+	-@test -d $(OBJDIR) && rm -fr $(OBJDIR)
+	-@test -e $(PROJECT).jar && rm  $(PROJECT).jar
 
 manifest:
 	mkdir $(OBJDIR)/META-INF && echo "Main-Class: $(TARGET)" > $(OBJDIR)/META-INF/MANIFEST.MF
+
 jar: manifest
 	(cd $(OBJDIR) && jar cvmf META-INF/MANIFEST.MF ../$(PROJECT).jar $(TARGETDIR)/* )
